pyproject.toml
[project]
name = "meetingbot"
version = "0.1.0"
description = "AI meeting assistant with real-time transcription, M365 integration, and multilingual support"
requires-python = ">=3.12,<3.14"
dependencies = [
    # API server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Settings
    "pydantic-settings>=2.3.0",
    # Azure Speech
    "azure-cognitiveservices-speech>=1.40.0",
    # Azure OpenAI
    "openai>=1.50.0",
    # Azure Document Intelligence
    "azure-ai-documentintelligence>=1.0.0",
    # Azure AI Search
    "azure-search-documents>=11.6.0",
    # Azure Blob Storage
    "azure-storage-blob>=12.22.0",
    # Azure Cosmos DB
    "azure-cosmos>=4.7.0",
    # Microsoft Graph
    "msgraph-sdk>=1.5.0",
    "azure-identity>=1.17.0",
    # Bing Search
    "requests>=2.32.0",
    "httpx>=0.27.0",
    # Document generation
    "python-docx>=1.1.2",
    # Mic capture (local meeting runner)
    "sounddevice>=0.4.0",
    # Utilities
    "tiktoken>=0.7.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.6.0",
    "numpy>=1.26.0",  # for audio int16 conversion
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
